founderforge_env/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-founderforge_env"
version = "0.1.0"
description = "FounderForge Startup CEO Sim for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0"
]

[project.scripts]
server = "founderforge_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["founderforge_env", "founderforge_env.server"]
